# Conflicts:
#	build.gradle
#	core/src/main/assets/messages/windows/windows.properties
#	core/src/main/assets/messages/windows/windows_nl.properties
#	core/src/main/assets/messages/windows/windows_uk.properties
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/Char.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/Hero.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/Perks.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Slime.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/npcs/Blacksmith.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/artifacts/DriedRose.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfWealth.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/MagesStaff.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Rapier.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Spear.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/scenes/StartScene.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/ui/Button.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/ui/GameLog.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndBlacksmith.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndGameInProgress.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndJournal.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndSettings.java
#	gradle/wrapper/gradle-wrapper.properties

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Momentum.java
@@ -3,10 +3,10 @@
  * Copyright (C) 2012-2015 Oleg Dolya
  *
  * Shattered Pixel Dungeon
- * Copyright (C) 2014-2023 Evan Debenham
+ * Copyright (C) 2014-2024 Evan Debenham
  *
  * Experienced Pixel Dungeon
- * Copyright (C) 2019-2020 Trashbox Bobylev
+ * Copyright (C) 2019-2024 Trashbox Bobylev
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -120,9 +120,9 @@ public class Momentum extends Buff implements ActionIndicator.Action {
 		}
 	}
 	
-	public int evasionBonus( int heroLvl, int excessArmorStr ){
+	public int evasionBonus(long heroLvl, long excessArmorStr ){
 		if (freerunTurns > 0) {
-			return heroLvl/2 + excessArmorStr*Dungeon.hero.pointsInTalent(Talent.EVASIVE_ARMOR);
+			return (int) (heroLvl/2 + excessArmorStr*Dungeon.hero.pointsInTalent(Talent.EVASIVE_ARMOR));
 		} else {
 			return 0;
 		}
